+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/DefaultEnergyUsage.java
@@ -34,6 +34,9 @@ public final class DefaultEnergyUsage {
     public static final long FORTUNE_3_UPGRADE = 14;
     public static final long SILK_TOUCH_UPGRADE = 16;
     public static final long REGULATOR_UPGRADE = 16;
+    public static final long RANGE_UPGRADE = 8;
+    public static final long CREATIVE_RANGE_UPGRADE = 0;
+    public static final int RANGE_UPGRADE_RANGE = 8;
 
     public static final long WIRELESS_GRID_CAPACITY = 1000;
     public static final long WIRELESS_GRID_OPEN = 5;