+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/ConfigImpl.java
@@ -543,6 +543,9 @@ public class ConfigImpl implements Config {
         private final ForgeConfigSpec.LongValue fortune3UpgradeEnergyUsage;
         private final ForgeConfigSpec.LongValue silkTouchUpgradeEnergyUsage;
         private final ForgeConfigSpec.LongValue regulatorUpgradeEnergyUsage;
+        private final ForgeConfigSpec.LongValue rangeUpgradeEnergyUsage;
+        private final ForgeConfigSpec.LongValue creativeRangeUpgradeEnergyUsage;
+        private final ForgeConfigSpec.IntValue rangeUpgradeRange;
 
         UpgradeEntryImpl() {
             builder.push("upgrade");
@@ -567,6 +570,20 @@ public class ConfigImpl implements Config {
             regulatorUpgradeEnergyUsage = builder
                 .comment("The additional energy used by the Regulator Upgrade")
                 .defineInRange("regulatorUpgradeEnergyUsage", DefaultEnergyUsage.REGULATOR_UPGRADE, 0, Long.MAX_VALUE);
+            rangeUpgradeEnergyUsage = builder
+                .comment("The additional energy used by the Range Upgrade")
+                .defineInRange("rangeUpgradeEnergyUsage", DefaultEnergyUsage.RANGE_UPGRADE, 0, Long.MAX_VALUE);
+            creativeRangeUpgradeEnergyUsage = builder
+                .comment("The additional energy used by the Creative Range Upgrade")
+                .defineInRange(
+                    "creativeRangeUpgradeEnergyUsage",
+                    DefaultEnergyUsage.CREATIVE_RANGE_UPGRADE,
+                    0,
+                    Long.MAX_VALUE
+                );
+            rangeUpgradeRange = builder
+                .comment("The additional range by the Range Upgrade")
+                .defineInRange("rangeUpgradeRange", DefaultEnergyUsage.RANGE_UPGRADE_RANGE, 0, Integer.MAX_VALUE);
             builder.pop();
         }
 
@@ -603,6 +620,21 @@ public class ConfigImpl implements Config {
         @Override
         public long getRegulatorUpgradeEnergyUsage() {
             return regulatorUpgradeEnergyUsage.get();
+        }
+
+        @Override
+        public long getRangeUpgradeEnergyUsage() {
+            return rangeUpgradeEnergyUsage.get();
+        }
+
+        @Override
+        public long getCreativeRangeUpgradeEnergyUsage() {
+            return creativeRangeUpgradeEnergyUsage.get();
+        }
+
+        @Override
+        public int getRangeUpgradeRange() {
+            return rangeUpgradeRange.get();
         }
     }