+++ CHANGELOG.md
@@ -12,6 +12,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Wireless Grid
 -   Creative Wireless Grid
 -   Wireless Transmitter
+-   Range Upgrade
+-   Creative Range Upgrade
 
 ### Changed